test(stella-tui): split the GRAPH query-bar goldens into their own module
CI's file-size ratchet went red on the merge, not on either side: this
branch added 63 lines to `deck_render_snapshots.rs` and #4780 added 63 more
after the last merge here, composing to 1501 — one over the limit. Neither
tree is wrong on its own, which is the shared-cell shape the ratchet's own
docs describe.

Split rather than a baseline entry, which the guard refuses and AGENTS.md
forbids for a first-time crossing. The three GRAPH bar goldens — a measured
query time, a neighborhood answering a free-form query, and the query box
mid-type — move to `tests/deck_render_snapshots/graph.rs`, sharing every
helper through `use super::*` and blessing with the same command. The
parent drops to 1359 lines, so it has room for the next tab that wants a row.

`#[path]` because this target's module root is `tests/`: a plain `mod graph`
would want `tests/graph.rs`, and Cargo compiles every `tests/*.rs` as its own
test binary, which would build these three goldens twice under two names.

No test is deleted or renamed; three move file.

Also repairs a comment this branch made stale: `attachment_block`'s
audio/video arm in `anthropic.rs` said `wire_parts` had "already degraded
them to Text notes", which is now only true of audio — video degrades to
sampled Image parts plus the note saying they are frames.
