+++ crates/stella-tui/tests/deck_render_snapshots.rs
@@ -68,6 +68,18 @@ use stella_tui::{
     render_deck,
 };
 
+/// The GRAPH tab's query-bar goldens. Split out because this file is the
+/// whole deck's golden surface and every tab wants rows in it, so it is the
+/// one that reaches the 1500-line ceiling first; the submodule shares every
+/// helper below and blesses with the same command.
+///
+/// `#[path]` because this target's module root is `tests/`, so a plain
+/// `mod graph` would want `tests/graph.rs` — and Cargo compiles every
+/// `tests/*.rs` as its own test binary, which would build these three
+/// goldens twice under two names.
+#[path = "deck_render_snapshots/graph.rs"]
+mod graph;
+
 /// The command used to regenerate every golden in this file. Quoted verbatim in
 /// each snapshot header and in every failure message, so nobody has to go
 /// looking for it.
@@ -670,85 +682,6 @@ fn deck_render_snapshots_pin_a_multiline_successful_tool_result() {
     assert_golden(
         "session_tool_result_multiline",
         "a folded multi-line success and a syntax-coloured JSON result",
-        W,
-        H,
-        &frame,
-    );
-}
-
-/// **The witness (#4335).** The GRAPH query bar reports what the query cost
-/// when the driver measured one.
-///
-/// A second golden rather than a changed `tab_graph`: the demo snapshot is
-/// synthesized and carries no timing, and that is the state the bar must keep
-/// rendering — `0ms` on a query nobody ran would be worse than silence. So
-/// both halves are pinned, the untimed one by `tab_graph` and the timed one
-/// here.
-#[test]
-fn deck_render_snapshots_pin_the_graph_query_time() {
-    let model = fixture_model();
-    let mut ui = ui_for(DeckTab::Graph);
-    if let Some(graph) = ui.graph.as_mut() {
-        graph.query_ms = Some(12);
-    }
-    let frame = render_frame(&model, &mut ui, W, H);
-    assert!(
-        frame.contains("· 12ms ·"),
-        "the query bar reports the timing:\n{frame}"
-    );
-    assert_golden(
-        "tab_graph_timed",
-        "the GRAPH tab with a measured query time in the query bar",
-        W,
-        H,
-        &frame,
-    );
-}
-
-/// **The witness (#4335).** The GRAPH query bar's second mode: a
-/// neighborhood that answers a free-form query reads `q:<text>`, not
-/// `file:<focus>`, so it names which of the two ways of re-rooting produced
-/// what is on screen.
-#[test]
-fn deck_render_snapshots_pin_the_graph_free_form_query() {
-    let model = fixture_model();
-    let mut ui = ui_for(DeckTab::Graph);
-    if let Some(graph) = ui.graph.as_mut() {
-        graph.query = Some("run_turn".into());
-        graph.query_ms = Some(12);
-    }
-    let frame = render_frame(&model, &mut ui, W, H);
-    assert!(
-        frame.contains("q:run_turn"),
-        "the bar reads the query it answers:\n{frame}"
-    );
-    assert!(
-        !frame.contains("file:"),
-        "and not the file selector it is standing in for:\n{frame}"
-    );
-    assert_golden(
-        "tab_graph_query",
-        "the GRAPH tab rooted on a free-form query rather than a file",
-        W,
-        H,
-        &frame,
-    );
-}
-
-/// The query box while it is being typed: the bar echoes the buffer with a
-/// caret, and the footer swaps the tab's keys for the box's, because those
-/// are the only ones that do anything while it is up.
-#[test]
-fn deck_render_snapshots_pin_the_graph_query_box() {
-    let model = fixture_model();
-    let mut ui = ui_for(DeckTab::Graph);
-    ui.graph_query = Some("run_tu".into());
-    let frame = render_frame(&model, &mut ui, W, H);
-    assert!(frame.contains("q:run_tu"), "{frame}");
-    assert!(frame.contains("run query"), "{frame}");
-    assert_golden(
-        "tab_graph_query_box",
-        "the GRAPH tab with the free-form query box open mid-type",
         W,
         H,
         &frame,
